sl-98: + UsePasswordForm hook
