+++ frontend/src/libs/hooks/hooks.ts
@@ -1,5 +1,6 @@
 export { useAppDispatch } from './use-app-dispatch/use-app-dispatch.hook.js';
 export { useAppForm } from './use-app-form/use-app-form.hook.js';
+export { UsePasswordForm } from './use-app-password-form/use-app-password-form.js';
 export { useAppSelector } from './use-app-selector/use-app-selector.hook.js';
 export { useStepper } from './use-app-stepper/use-app-stepper.hook.js';
 export { useCallback, useEffect, useMemo, useRef, useState } from 'react';
