+++ frontend/src/pages/account-settings/components/login/styles.module.scss
@@ -19,17 +19,3 @@
   top: 40%;
   right: 3%;
 }
-
-.buttons {
-  display: flex;
-  justify-content: space-between;
-  gap: 13px;
-}
-
-.button {
-  width: 100%;
-}
-
-.submit-button {
-  color: var(--black);
-}
